requirements.txt
beautifulsoup4==4.4.1
Django==1.8.4
django-htmlmin==0.8.0
django-registration-redux==1.2
huey==0.4.9
lxml==3.4.4
psycopg2==2.6.1
pylibmc==1.5.0
redis==2.10.5
uWSGI==2.0.11.2
